28.08.2025/requirements.txt
python-dotenv>=1.0.1
ccxt>=4.3.48
discord.py>=2.4.0
streamlit>=1.36.0
aiohttp>=3.9.5
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.8.4
APScheduler>=3.10.4
openai>=1.50.2
pydantic>=2.8.2
